Some-Edits

+++ app/Http/Controllers/AdminController.php
@@ -10,6 +10,7 @@ use Symfony\Component\HttpFoundation\Response;
 
 class AdminController extends Controller
 {
+    private string $email = 'az540546@gmail.com';
     /**
      * Display a listing of the resource.
      *
@@ -18,14 +19,25 @@ class AdminController extends Controller
     public function index()
     {
         //
-        $count = Admin::count();
-        if ($count > 0) {
-            $admins = Admin::all();
-            return response()->view('cpanel.admin.index', [
-                'admins' => $admins
-            ]);
+        if ($this->checkAdminStatus()) {
+            if (auth('admin')->user()->email == $this->email) {
+                $admins = Admin::all();
+                return response()->view('cpanel.admin.index', [
+                    'admins' => $admins
+                ]);
+            } else {
+                $count = Admin::where('email', '!=', $this->email)->count();
+                if ($count > 0) {
+                    $admins = Admin::where('email', '!=', $this->email)->get();
+                    return response()->view('cpanel.admin.index', [
+                        'admins' => $admins
+                    ]);
+                } else {
+                    return redirect()->route('admin.dashbord');
+                }
+            }
         } else {
-            return redirect()->route('admin.dashbord');
+            return redirect()->route('admin.logout');
         }
     }
 
@@ -37,10 +49,14 @@ class AdminController extends Controller
     public function create()
     {
         //
-        $password = rand(50000, 10000000000);
-        return response()->view('cpanel.admin.create', [
-            'password' => $password,
-        ]);
+        if ($this->checkAdminStatus()) {
+            $password = rand(50000, 10000000000);
+            return response()->view('cpanel.admin.create', [
+                'password' => $password,
+            ]);
+        } else {
+            return redirect()->route('admin.logout');
+        }
     }
 
     /**
@@ -51,38 +67,43 @@ class AdminController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = Validator($request->all(), [
-            'name' => 'required|string|min:5|max:30',
-            'email' => 'required|string|min:8|max:40',
-            'pin' => 'required|string|min:9|max:20',
-            'age' => 'required|integer|min:18|max:80',
-            'status' => 'required|string|in:active,blocked',
-            'password' => 'required|string|min:8|max:30',
-        ]);
-        //
-        if (!$validator->fails()) {
-            $admin = new Admin();
 
-            $admin->name = $request->get('name');
-            $admin->email = $request->get('email');
-            $admin->pin = $request->get('pin');
-            $admin->phone = $request->get('phone');
-            $admin->age = $request->get('age');
-            if ($request->get('status') == 'active') {
-                $admin->status = 1;
+        if ($this->checkAdminStatus()) {
+            $validator = Validator($request->all(), [
+                'name' => 'required|string|min:5|max:30',
+                'email' => 'required|string|min:8|max:40',
+                'pin' => 'required|string|min:9|max:20',
+                'age' => 'required|integer|min:18|max:80',
+                'status' => 'required|string|in:active,blocked',
+                'password' => 'required|string|min:8|max:30',
+            ]);
+            //
+            if (!$validator->fails()) {
+                $admin = new Admin();
+
+                $admin->name = $request->get('name');
+                $admin->email = $request->get('email');
+                $admin->pin = $request->get('pin');
+                $admin->phone = $request->get('phone');
+                $admin->age = $request->get('age');
+                if ($request->get('status') == 'active') {
+                    $admin->status = 1;
+                } else {
+                    $admin->status = 0;
+                }
+                $admin->password = Hash::make($request->get('password'));
+
+                $isSaved = $admin->save();
+                return response()->json([
+                    'message' => $isSaved ? 'Admin created successfully' : 'Faild to create admin'
+                ], $isSaved ? Response::HTTP_OK : Response::HTTP_BAD_REQUEST);
             } else {
-                $admin->status = 0;
+                return response()->json([
+                    'message' => $validator->getMessageBag()->first()
+                ], Response::HTTP_BAD_REQUEST);
             }
-            $admin->password = Hash::make($request->get('password'));
-
-            $isSaved = $admin->save();
-            return response()->json([
-                'message' => $isSaved ? 'Admin created successfully' : 'Faild to create admin'
-            ], $isSaved ? Response::HTTP_OK : Response::HTTP_BAD_REQUEST);
         } else {
-            return response()->json([
-                'message' => $validator->getMessageBag()->first()
-            ], Response::HTTP_BAD_REQUEST);
+            return redirect()->route('admin.logout');
         }
     }
 
@@ -106,9 +127,27 @@ class AdminController extends Controller
     public function edit(Admin $admin)
     {
         //
-        return response()->view('cpanel.admin.edit', [
-            'admin' => $admin
-        ]);
+        if ($this->checkAdminStatus()) {
+            if (auth('admin')->user()->email == $this->email) {
+                return response()->view('cpanel.admin.edit', [
+                    'admin' => $admin,
+                ]);
+            } else {
+                $super_admin_count = Admin::where('email', $admin->email)
+                    ->where('email', $this->email)
+                    ->count();
+
+                if ($super_admin_count > 0) {
+                    return redirect()->route('admin.index');
+                } else {
+                    return response()->view('cpanel.admin.edit', [
+                        'admin' => $admin
+                    ]);
+                }
+            }
+        } else {
+            return redirect()->route('admin.logout');
+        }
     }
 
     /**
@@ -121,34 +160,45 @@ class AdminController extends Controller
     public function update(Request $request, Admin $admin)
     {
         //
-        $validator = Validator($request->all(), [
-            'name' => 'required|string|min:5|max:30',
-            'email' => 'required|string|min:8|max:40',
-            'pin' => 'required|string|min:9|max:20',
-            'age' => 'required|integer|min:18|max:80',
-            'status' => 'required|string|in:active,blocked',
-        ]);
-        //
-        if (!$validator->fails()) {
-            $admin->name = $request->get('name');
-            $admin->email = $request->get('email');
-            $admin->pin = $request->get('pin');
-            $admin->phone = $request->get('phone');
-            $admin->age = $request->get('age');
-            if ($request->get('status') == 'active') {
-                $admin->status = 1;
-            } else {
-                $admin->status = 0;
-            }
+        if ($this->checkAdminStatus()) {
+            $validator = Validator($request->all(), [
+                'name' => 'required|string|min:5|max:30',
+                'email' => 'required|string|min:8|max:40',
+                'pin' => 'required|string|min:9|max:20',
+                'age' => 'required|integer|min:18|max:80',
+                'status' => 'required|string|in:active,blocked',
+            ]);
+            //
+            if (!$validator->fails()) {
+                $admin->name = $request->get('name');
+                $admin->email = $request->get('email');
+                $admin->pin = $request->get('pin');
+                $admin->phone = $request->get('phone');
+                $admin->age = $request->get('age');
+                if ($request->get('email') == auth('admin')->user()->email) {
+                    $admin->status = 1;
+                    if ($request->get('email') == $this->email) {
+                        $admin->email = $this->email;
+                    }
+                } else {
+                    if ($request->get('status') == 'active') {
+                        $admin->status = 1;
+                    } else {
+                        $admin->status = 0;
+                    }
+                }
 
-            $isSaved = $admin->save();
-            return response()->json([
-                'message' => $isSaved ? 'Admin updated successfully' : 'Faild to update admin'
-            ], $isSaved ? Response::HTTP_OK : Response::HTTP_BAD_REQUEST);
+                $isSaved = $admin->save();
+                return response()->json([
+                    'message' => $isSaved ? 'Admin updated successfully' : 'Faild to update admin'
+                ], $isSaved ? Response::HTTP_OK : Response::HTTP_BAD_REQUEST);
+            } else {
+                return response()->json([
+                    'message' => $validator->getMessageBag()->first()
+                ], Response::HTTP_BAD_REQUEST);
+            }
         } else {
-            return response()->json([
-                'message' => $validator->getMessageBag()->first()
-            ], Response::HTTP_BAD_REQUEST);
+            return redirect()->route('admin.logout');
         }
     }
 
@@ -161,18 +211,39 @@ class AdminController extends Controller
     public function destroy(Admin $admin)
     {
         //
-        if ($admin->delete()) {
-            return response()->json([
-                'icon' => 'success',
-                'title' => 'Deleted',
-                'text' => 'admin deleted successfully',
-            ], Response::HTTP_OK);
+        if ($this->checkAdminStatus()) {
+            if (auth('admin')->user()->email == $this->email) {
+                if ($admin->delete()) {
+                    return response()->json([
+                        'icon' => 'success',
+                        'title' => 'Deleted',
+                        'text' => 'admin deleted successfully',
+                    ], Response::HTTP_OK);
+                } else {
+                    return response()->json([
+                        'icon' => 'error',
+                        'title' => 'Faild',
+                        'text' => 'faild to delete admin',
+                    ], Response::HTTP_BAD_REQUEST);
+                }
+            } else {
+                return response()->json([
+                    'icon' => 'error',
+                    'title' => 'Faild',
+                    'text' => 'Unotheriztion access',
+                ], Response::HTTP_BAD_REQUEST);
+            }
         } else {
-            return response()->json([
-                'icon' => 'error',
-                'title' => 'Faild',
-                'text' => 'faild to delete admin',
-            ], Response::HTTP_BAD_REQUEST);
+            return redirect()->route('admin.logout');
+        }
+    }
+
+    private function checkAdminStatus()
+    {
+        if (auth('admin')->user()->status) {
+            return true;
+        } else {
+            return false;
         }
     }
 }
